Revert "tryit" and instead rename packages proper
This reverts commit 02cfb62cbe977019c8b3d56d71fc634d9aa6b742.
