+++ src/main/java/quiz/model/Answer.java
@@ -1,4 +1,4 @@
-package see.hm.edu.quizmodel;
+package quiz.model;
 
 public class Answer {
 	public String text;
