+++ src/main/java/quiz/model/Player.java
@@ -1,4 +1,4 @@
-package see.hm.edu.quizmodel;
+package quiz.model;
 
 public class Player {
 	public int score;
