+++ src/main/java/quiz/model/Question.java
@@ -1,4 +1,4 @@
-package see.hm.edu.quizmodel;
+package quiz.model;
 
 import java.net.URL;
 import java.util.Arrays;
